requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
pillow==10.3.0
# mysqlclient==2.2.4
python-decouple==3.8
django_jazzmin==3.0.0
django-cron==0.6.0
django-crontab==0.7.1
pytz
python-dateutil